Add an opt-in tf32 path for the QR trailing updates; up to 1.9x
Profiling showed the matmuls at ~70% of GPU time, and within them only the
trailing-update GEMMs matter. Forcing strategy_tall to Strategy::Auto takes
1024x1024 from 23.5ms to 12.0ms, while forcing the Gram or W matmul to Auto
changes nothing end to end. Auto resolves to a tensor-core path that rounds
f32 inputs to tf32, which is why it was ruled out wholesale before.

Expose it as a knob instead of leaving the speed on the table:

- BahtTsqrStrategy gains `allow_tf32` (off by default), mapped into the
  blueprint and applied to strategy_tall only. Gram and W stay
  full-precision unconditionally since they cost nothing either way.
- New `qr_with_strategy()` entry point; `qr()` delegates with defaults, so
  the existing API and its accuracy are unchanged.
- profile_qr takes an optional 5th arg `tf32` so both paths are profilable.

The flag is gated on `supports_type(tf32)`. Without a tensor-core path
Strategy::Auto resolves to something *slower* than the full-precision unit
routine -- measured on wgpu, 24.1ms -> 30.5ms at 1024x1024 -- so honouring
the opt-in there would cost accuracy and speed. Post-gate wgpu measures
21.30/21.56/21.35ms with the flag against 21.30/21.24ms without: a genuine
no-op. Also a no-op for f64 (112.7 vs 112.4ms), which has no tf32 path.

CUDA f32 medians (RTX 4070 Laptop), full -> tf32:
  128x128:     1.34ms ->  1.26ms  (1.06x)
  512x512:     7.63ms ->  5.31ms  (1.44x)
  1024x1024:  22.84ms -> 11.90ms  (1.92x)
  2048x512:   22.01ms -> 11.40ms  (1.93x)
  2048x2048:  80.22ms -> 48.68ms  (1.65x)

The gain scales with size -- small problems are dominated by panel dispatch
overhead, not matmul -- so the flag is worth enabling for large
factorizations and roughly pointless below ~256.

Gap to PyTorch/cuSOLVER f32 on the same GPU narrows from 2.2-6.8x to
2.1-4.1x. (cuSOLVER uses tensor cores for f32 too, so the tf32 column is
the closer like-for-like comparison.)

Tests: adds test_tf32_opt_in (f32 and f64) driving qr_with_strategy at
157x157, asserted at 5e-2 -- the tolerance matching the documented ~1e-2
accuracy rather than pretending it is f32-accurate. 22/22 pass on CUDA and
wgpu. Also fixes two unnecessary-cast clippy warnings introduced by the
earlier vec_split helper.

+++ crates/cubek-linalg/src/components/baht_tsqr.rs
@@ -334,10 +334,10 @@ pub fn launch<R: Runtime, E: Float + CubeElement>(
     let vec_split = |n: usize| -> (VectorSize, usize) {
         let v = client
             .io_optimized_vector_sizes(elem_size)
-            .filter(|v| n.is_multiple_of(*v as usize))
+            .filter(|v| n.is_multiple_of(*v))
             .max()
             .unwrap_or(1);
-        (v, n / v as usize)
+        (v, n / v)
     };
 
     let launch_matmul = |strategy: &Strategy,
